fixed normnette/organized files/added fix for $ standalone
