+++ Makefile
@@ -27,9 +27,10 @@ C_FILES = debug.c \
 	tokenizer/tokenizer_utils.c \
 	tokenizer/token_utils.c \
 	tokenizer/quote_handler.c \
-	vars_expansion/vars_exp_init.c \
-	vars_expansion/var_expansion.c \
-	vars_expansion/utils.c \
+	vars_expansion/exp_init.c \
+	vars_expansion/var_name_extraction.c \
+	vars_expansion/special_cases.c \
+	vars_expansion/expansion.c \
 	parser/cmd/init_list.c \
 	parser/cmd/new_cmd.c \
 	parser/cmd/utils.c \
